requirements.txt
asgiref==3.7.2
# backports.zoneinfobackports.zoneinfo;python_version<"3.9"
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
gunicorn==21.2.0
importlib-metadata==6.8.0
Markdown==3.4.4
python-dotenv==1.0.0
pytz==2022.7
six==1.16.0
sqlparse==0.4.4
zipp==3.16.2
lxml==4.9.3